Move P2 modes parsers
