+++ src/consts.rs
@@ -1,3 +1,5 @@
+use crate::AppSW;
+
 pub const ZCASH_TICKER: &str = "ZEC";
 
 pub const ZCASH_DECIMALS: u32 = 8;
@@ -37,3 +39,41 @@ pub const P2_FINALIZE_FULL_DEFAULT: u8 = 0x00;
 
 pub const TRUSTED_INPUT_SIZE: usize = 2 + 2 + 32 + 4 + 8; // magic + rand + txid + idx + amount
 pub const TRUSTED_INPUT_TOTAL_SIZE: usize = TRUSTED_INPUT_SIZE + 8;
+
+#[derive(Debug, Clone, Copy)]
+#[repr(u8)]
+pub enum P2VkMode {
+    Ufvk = 0x0,
+    OrchardFvk = 0x1,
+}
+
+impl TryFrom<u8> for P2VkMode {
+    type Error = AppSW;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0x00 => Ok(P2VkMode::Ufvk),
+            0x01 => Ok(P2VkMode::OrchardFvk),
+            _ => Err(AppSW::WrongP1P2),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+#[repr(u8)]
+pub enum P2ShieldedAddrMode {
+    UAddress = 0x0,
+    OrchardAddress = 0x1,
+}
+
+impl TryFrom<u8> for P2ShieldedAddrMode {
+    type Error = AppSW;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0x00 => Ok(P2ShieldedAddrMode::UAddress),
+            0x01 => Ok(P2ShieldedAddrMode::OrchardAddress),
+            _ => Err(AppSW::WrongP1P2),
+        }
+    }
+}
